# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Product Inventory",
    "description": "API for managing products, categories, and users",
    "version": "1.0.0"
  },
  "host": "productinventory-90te.onrender.com",
  "basePath": "/",
  "schemes": ["https", "http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "description": "Get user lists",
        "security": [{ "oauth2": ["read:user"] }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Users"],
        "description": "Create a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "description": "Get a user by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "security": [{ "oauth2": ["read:user"] }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Users"],
        "description": "Update a user by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "description": "Delete a user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
...
    },
    "definitions": {
...
      "User": {
        "type": "object",
        "required": ["firstName", "lastName", "email"],
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name",
            "example": "Alejandro"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "User's email address",
            "example": "avillamar@example.com"
          },
          "favoriteColor": {
            "type": "string",
            "description": "User's favorite color",
            "example": "blue"
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "description": "User's birth date",
            "example": "1990-01-01"
          }
        }
      },
...
    },
    "securityDefinitions": {
      "oauth2": {
        "type": "oauth2",
        "flow": "accessCode",
        "authorizationUrl": "https://github.com/login/oauth/authorize",
        "tokenUrl": "https://github.com/login/oauth/access_token",
        "scopes": {
          "read:user": "Read user information"
        }
      }
    },
    "security": [
      {
        "oauth2": ["read:user"]
      }
    ]
}